cf_cleaner_tweet/requirements.txt
stop_words==2018.7.23
google-cloud-bigquery==1.20.0
google-cloud-bigquery-storage==0.7.0
pandas==0.25.1
pandas-gbq==0.11.0
langdetect==1.0.7
textblob==0.15.3